+++ source/common/tables/dbo.appConfigurations.sql
@@ -75,6 +75,7 @@ INSERT	INTO [dbo].[appConfigurations] ([module], [name], [value])
 		  SELECT 'maintenance-plan'	AS [module], 'WAIT_AT_LOW_PRIORITY max duration (min)'										AS [name], '1'			AS [value]		UNION ALL
 		  SELECT 'maintenance-plan'	AS [module], 'Use Default Scheduler for maintenance tasks if project specific not defined'	AS [name], '1'			AS [value]		UNION ALL
 		  SELECT 'maintenance-plan' AS [module], 'SMART default changes threshold'												AS [name], '50'			AS [value]		UNION ALL
+		  SELECT 'maintenance-plan' AS [module], 'Do not run maintenance for indexes not used in the last N days (0=disabled)'	AS [name], '30'			AS [value]		UNION ALL
 		
 
 		  SELECT 'health-check'		AS [module], 'Collect SQL Agent jobs step details'											AS [name], 'false'		AS [value]		UNION ALL